endianess: now the library is endian agnostic
Observations:

The only code that required little or big-endian was the multi-byte CRC32
optimization (i.e. it was not compatible with e.g. PDP endianess). Now
that optimization is disabled when the target CPU is not little-endian
nor big-endian. Remaining cases using endianess detection are for
optimizations for little-endian load/store operations (used in the data
compression functions), so non little-endian platforms are a bit slower
in that regard.