test_ss_cat hardening: multiple aliasing input
Observations:
- With this change the test targets to ensure the inputs being
concatenated respect the input string idempotency, i.e. when operating
with input references that are also the output, ensure the input is
processed like a constant.
- "Aliasing" means in this context being the input and the output the
same container, e.g. concatenating a string to itself

Example:

	ss_t *s = ss_dup_c("a");
	const ss_t *t = ss_crefa("b");
	ss_cat(&s, t, s, t, s, t, s, t);

Expected result for 's' would be "abababab", because of 's' considered as
an input constant. Internally we just keep the input string lengths, so we
can modify the output directly, without intermediate copies (i.e. the
operation is always efficient, no matter if there is aliasing or not).