data compression enhancements
Allow multiple LUTs instead of one for data compression pattern matching,
so output size can be reduced more. Default is kept in one LUT because of
CPU usage (the LUT minimum amount is 1, and the maximum, 40; the "z" mode
is limited to 2 LUTs, and the "zh" to 40). For general purpose it is not
recommended using more than 4 LUTs (the data compression gain gets smaller
with every LUT you add, so after a given amount of LUTs the gain could be
negligible).

E.g. compilation flag for enabling up to 4 LUTs: S_LZ_LUTS=4