O(n) string search complexity set as default
String search starts with a very fast quasi-O(n) algorithm (Rabin-Karp
search algorithm with a very fast hash function), and switches to a
slower but with de facto O(n) algorithm (Rabin-Karp, too, but with a more
robust hash function). I.e. with a small penalty (<5%) for the fast-slow
algorithm switching detection, the high performance is kept (the cases
where the switch to the "slow" algorithm happen are rare in real-world
usage, being hard even trying to do it "on purpose"). The "fast"
algorithm is 2 times faster than the "slow" (2GB/s vs 1GB/s on 3GHz
OooE CPU).